+++ src/ship/domain/universe/TimeDependent.java
@@ -13,7 +13,7 @@ public interface TimeDependent {
 
     public void beforeTimeElapsed();
 
-    public void timeElapsed(double timeElapsed);
+    public void timeElapsed(long timeElapsed);
 
     public void afterTimeElapsed();
